moved markov_utilities out of analysis
